4251: cleanup, split CaseInformationInternal.jsx because it was too long and exceed 700 line limit, leaving integration tests WIP for dod
